+++ src/index.js
@@ -63,9 +63,9 @@ const vbexp = function(configPath, overrideOptions) {
 		app._onStarting()
 
 		// Get port from environment and store in Express.
-		let host = app.config.get('server:host')
-		let port = app.config.get('server:port')
-		let localpath = app.config.get('server:path')
+		let host = app.config.get('server.host')
+		let port = app.config.get('server.port')
+		let localpath = app.config.get('server.path')
 
 		debug('host', host)
 		debug('port', port)